# Tritium provides an opt-in flag to use ByteBuddy instead of dynamic proxies