--- a/tritium-lib/build.gradle
+++ b/tritium-lib/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     api project(':tritium-tracing')
     api 'io.dropwizard.metrics:metrics-core'
 
+    implementation 'net.bytebuddy:byte-buddy'
     implementation 'com.google.code.findbugs:jsr305'
     implementation 'com.google.errorprone:error_prone_annotations'
     implementation 'com.google.guava:guava'