--- a/versions.props
+++ b/versions.props
@@ -15,6 +15,7 @@ com.palantir.tracing:* = 3.1.0
 com.squareup.okhttp3:okhttp = 4.0.0
 com.uber.nullaway:nullaway = 0.7.6
 io.undertow:undertow-core = 2.0.20.Final
+net.bytebuddy:byte-buddy = 1.10.1
 org.assertj:assertj-core = 3.13.2
 org.awaitility:awaitility = 4.0.0
 org.checkerframework:checker-qual = 2.5.0