… changelog/0.15.4-rc1/pr-378.v2.yml
… tritium-lib/src/main/java/com/palantir/tritium/proxy/ByteBuddyInstrumentationAdvice.java
… tritium-lib/src/main/java/com/palantir/tritium/proxy/Instrumentation.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/ByteBuddyInstrumentationTest.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/DynamicProxyInstrumentationTest.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/InstrumentationTest.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/subpackage/ByteBuddyInstrumentationPackageAccessTest.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/subpackage/PackagePrivateClass.java
… tritium-lib/src/test/java/com/palantir/tritium/proxy/subpackage/PackagePrivateInterface.java
… tritium-proxy/src/test/java/com/palantir/tritium/proxy/ProxiesTest.java
… tritium-test/src/main/java/com/palantir/tritium/test/LessSpecificReturn.java
… tritium-test/src/main/java/com/palantir/tritium/test/MoreSpecificReturn.java
… tritium-test/src/main/java/com/palantir/tritium/test/TestImplementation.java
… versions.lock