==COMMIT_MSG==
Tritium provides an opt-in flag to use ByteBuddy instead of dynamic proxies
==COMMIT_MSG==


If the `instrument.dynamic-proxy` property is set `false`, byte-buddy is used.

before:
```
java.lang.RuntimeException: Stack Trace
        at com.palantir.DemoMain.run(DemoMain.java:32)
        at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
        at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
        at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
        at java.base/java.lang.reflect.Method.invoke(Method.java:567)
        at com.palantir.tritium.proxy.InvocationEventProxy.invoke(InvocationEventProxy.java:102)
        at com.sun.proxy.$Proxy0.run(Unknown Source)
        at com.palantir.DemoMain.main(DemoMain.java:27)
```

after:
```
java.lang.RuntimeException: Stack Trace
        at com.palantir.DemoMain.run(DemoMain.java:32)
        at com.palantir.tritium.proxy.InstrumentedRunnable$0.run(Unknown Source)
        at com.palantir.DemoMain.main(DemoMain.java:27)
```

Notable differences:

* The bytebuddy implementation cannot instrument inaccessible interfaces, so it will log a warning and return the uninstrumented delegate object.
* Added an implementation dependency on byte-buddy, increasing the footprint of tritium-lib.

Benchmark results (haven't changed much, but they weren't the goal):
```
Benchmark                                                 (mode)  Mode  Cnt    Score    Error  Units
ProxyBenchmark.instrumentedDefaultTagged              BYTE_BUDDY  avgt    5  401.312 ± 17.375  ns/op
ProxyBenchmark.instrumentedDefaultTagged           DYNAMIC_PROXY  avgt    5  408.340 ± 10.732  ns/op
ProxyBenchmark.instrumentedDefaultUntagged            BYTE_BUDDY  avgt    5  518.907 ± 39.311  ns/op
ProxyBenchmark.instrumentedDefaultUntagged         DYNAMIC_PROXY  avgt    5  508.477 ±  7.854  ns/op
ProxyBenchmark.instrumentedWithEnabledNopHandler      BYTE_BUDDY  avgt    5   15.254 ±  0.457  ns/op
ProxyBenchmark.instrumentedWithEnabledNopHandler   DYNAMIC_PROXY  avgt    5   19.069 ±  0.555  ns/op
ProxyBenchmark.instrumentedWithEverything             BYTE_BUDDY  avgt    5  418.855 ± 57.766  ns/op
ProxyBenchmark.instrumentedWithEverything          DYNAMIC_PROXY  avgt    5  475.463 ± 46.280  ns/op
ProxyBenchmark.instrumentedWithMetrics                BYTE_BUDDY  avgt    5  285.950 ± 22.050  ns/op
ProxyBenchmark.instrumentedWithMetrics             DYNAMIC_PROXY  avgt    5  295.070 ± 15.405  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging     BYTE_BUDDY  avgt    5    7.195 ±  0.074  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging  DYNAMIC_PROXY  avgt    5   13.690 ±  0.344  ns/op
ProxyBenchmark.instrumentedWithTaggedMetrics          BYTE_BUDDY  avgt    5  214.352 ± 31.053  ns/op
ProxyBenchmark.instrumentedWithTaggedMetrics       DYNAMIC_PROXY  avgt    5  222.926 ± 62.158  ns/op
ProxyBenchmark.instrumentedWithTracing                BYTE_BUDDY  avgt    5  212.589 ±  6.832  ns/op
ProxyBenchmark.instrumentedWithTracing             DYNAMIC_PROXY  avgt    5  219.651 ±  6.300  ns/op
ProxyBenchmark.instrumentedWithTracingNested          BYTE_BUDDY  avgt    5  181.806 ±  3.045  ns/op
ProxyBenchmark.instrumentedWithTracingNested       DYNAMIC_PROXY  avgt    5  195.890 ±  3.228  ns/op
ProxyBenchmark.raw                                    BYTE_BUDDY  avgt    5    2.227 ±  0.095  ns/op
ProxyBenchmark.raw                                 DYNAMIC_PROXY  avgt    5    2.204 ±  0.058  ns/op
```